Before mkdir(2), check if directory exists.
On Cygwin, mkdir(2) reports EACCES if the directory exists but the
system call would have failed had it not existed. This is a little odd
since EEXIST would be a much more sensible error code, but this odd
behavior seems to be permitted by POSIX. If mkdir(2) fails, but the path
exists as a directory, then continue as normal.